Refactored company details and investment tab to move data parsing and labels into their own files and have the controller concentrate on being a controller. Change order of fields in company detail slightly and add and remove a few fields Fix a spacing issue in companies house table. Show message when no related companies
