+++ src/controllers/companycontroller.js
@@ -4,63 +4,11 @@ const winston = require('winston')
 const controllerUtils = require('../lib/controllerutils')
 const companyRepository = require('../repositorys/companyrepository')
 const itemCollectionService = require('../services/itemcollectionservice')
-const getFormattedAddress = require('../lib/address').getFormattedAddress
-const DateLib = require('../lib/date')
-const sentenceCase = require('../lib/sentencecase')
-
+const companyFormattingService = require('../services/companyformattingservice')
+const investmentFormattingService = require('../services/investmentformattingservice')
+const { companyDetailLabels, chDetailLabels, companyTableHeadings } = require('../labels/companylabels')
+const { investmentDetailLabels, investmentProjectsOpenLabels, investmentProjectsClosedLabels } = require('../labels/investmentlabels')
 const router = express.Router()
-const companyDetailLabels = {
-  registered_address: 'Registered address',
-  business_type: 'Company type',
-  alias: 'Trading name',
-  trading_address: 'Trading address',
-  uk_region: 'Region',
-  headquarters: 'Headquarters',
-  sector: 'Primary sector',
-  website: 'Website',
-  description: 'Business description',
-  employee_range: 'Number of employees',
-  turnover_range: 'Annual turnover'
-}
-const chDetailLabels = {
-  company_number: 'Companies House number',
-  registered_address: 'Registered office address',
-  business_type: 'Company type',
-  company_status: 'Company status',
-  sic_code: 'Nature of business (SIC)'
-}
-
-const companyDetailsDisplayOrder = Object.keys(companyDetailLabels)
-const chDetailsDisplayOrder = Object.keys(chDetailLabels)
-
-const investmentDetailLabels = {
-  account_management_tier: 'Account management tier',
-  account_manager: 'Account manager',
-  ownership: 'Ownership'
-}
-const investmentDetailsDisplayOrder = Object.keys(investmentDetailLabels)
-const investmentProjectsOpenLabels = {
-  name: 'Open projects',
-  value: 'Value',
-  state: 'Stage',
-  land_date: 'Land date'
-}
-const investmentProjectsOpenKeys = Object.keys(investmentProjectsOpenLabels)
-const investmentProjectsClosedLabels = {
-  name: 'Closed projects',
-  value: 'Value',
-  state: 'Status',
-  state_date: 'Status date'
-}
-const investmentProjectsClosedKeys = Object.keys(investmentProjectsClosedLabels)
-
-const TODO = '<span class="status-badge status-badge--xsmall status-badge--action">TO DO</span>'
-
-const companyTableHeadings = {
-  name: 'Company name',
-  address: 'Address'
-}
-const companyTableKeys = ['name', 'address']
 
 function cleanErrors (errors) {
   if (errors.registered_address_1 || errors.registered_address_2 ||
@@ -87,144 +35,6 @@ function cleanErrors (errors) {
     delete errors.trading_address_country
   }
 }
-function getDisplayCH (company) {
-  if (!company.companies_house_data) return null
-
-  const companyHouseData = company.companies_house_data
-
-  const displayCH = {
-    company_number: companyHouseData.company_number,
-    business_type: companyHouseData.company_category,
-    company_status: companyHouseData.company_status,
-    registered_address: getFormattedAddress(company.companies_house_data, 'registered')
-  }
-
-  displayCH.sic_code = []
-  if (companyHouseData.sic_code_1 && companyHouseData.sic_code_1.length > 0) displayCH.sic_code.push(companyHouseData.sic_code_1)
-  if (companyHouseData.sic_code_2 && companyHouseData.sic_code_1.length > 0) displayCH.sic_code.push(companyHouseData.sic_code_2)
-  if (companyHouseData.sic_code_3 && companyHouseData.sic_code_1.length > 0) displayCH.sic_code.push(companyHouseData.sic_code_3)
-  if (companyHouseData.sic_code_4 && companyHouseData.sic_code_1.length > 0) displayCH.sic_code.push(companyHouseData.sic_code_4)
-
-  return displayCH
-}
-function getDisplayCompany (company) {
-  if (!company.id) return null
-
-  const displayCompany = {
-    sector: (company.sector && company.sector.name) ? company.sector.name : TODO,
-    description: company.description || TODO,
-    website: company.website ? `<a href="${company.website}">${company.website}</a>` : TODO,
-    employee_range: (company.employee_range && company.employee_range.name) ? company.employee_range.name : TODO,
-    turnover_range: (company.turnover_range && company.turnover_range.name) ? company.turnover_range.name : TODO,
-    account_manager: (company.account_manager && company.account_manager.name) ? company.account_manager.name : TODO,
-    headquarters: 'UK headquarters'
-  }
-
-  if (company.alias && company.alias.length > 0) displayCompany.alias = company.alias
-
-  const tradingAddress = getFormattedAddress(company, 'trading')
-  if (tradingAddress.length > 0) displayCompany.trading_address = tradingAddress
-
-  if (!company.companies_house_data) {
-    displayCompany.registered_address = getFormattedAddress(company, 'registered')
-    displayCompany.business_type = (company.business_type && company.business_type.name) ? company.business_type.name : TODO
-  }
-
-  if (company.uk_region) displayCompany.uk_region = company.uk_region.name
-
-  if (company.export_to_countries && company.export_to_countries.length > 0) {
-    displayCompany.export_to_countries = company.export_to_countries.map(country => country.name).toString()
-  } else {
-    displayCompany.export_to_countries = 'No'
-  }
-  if (company.future_interest_countries && company.future_interest_countries.length > 0) {
-    displayCompany.future_interest_countries = company.future_interest_countries.map(country => country.name).toString()
-  } else {
-    displayCompany.future_interest_countries = 'No'
-  }
-
-  return displayCompany
-}
-function getHeadingAddress (company) {
-  // If this is a CDMS
-  const cdmsTradingAddress = getFormattedAddress(company, 'trading')
-  if (cdmsTradingAddress.length > 0) {
-    return cdmsTradingAddress
-  }
-
-  if (company.companies_house_data !== null) {
-    return getFormattedAddress(company.companies_house_data, 'registered')
-  }
-
-  return getFormattedAddress(company, 'registered')
-}
-function getHeadingName (company) {
-  if (company.id) {
-    if (company.alias && company.alias.length > 0) {
-      return company.alias
-    }
-    return company.name
-  } else {
-    return company.companies_house_data.name
-  }
-}
-function parseRelatedData (companies) {
-  return companies.map((company) => {
-    const key = (company.trading_address_1 && company.trading_address_1.length > 0) ? 'trading' : 'registered'
-
-    let address = ''
-    if (company[`${key}_address_town`] && company[`${key}_address_town`].length > 0) {
-      address += sentenceCase(`${company[`${key}_address_town`]}, `)
-    } else if (company[`${key}_address_county`] && company[`${key}_address_county`].length > 0) {
-      address += sentenceCase(`${company[`${key}_address_county`]}, `)
-    }
-    if (company[`${key}_address_country`] && company[`${key}_address_country`].name && company[`${key}_address_country`].name.length > 0) {
-      address += sentenceCase(company[`${key}_address_country`].name)
-    } else if (address.length > 0) {
-      address += 'United Kingdom'
-    }
-    return {
-      name: `<a href="/company/company_company/${company.id}">${company.alias || company.name}</a>`,
-      address
-    }
-  })
-}
-function getInvestmentDetailsDisplay (company) {
-  if (!company.id) return null
-  return {
-    account_management_tier: 'B - Top 300',
-    account_manager: `<a href="/advisor/${company.account_manager.id}/">${company.account_manager.name}</a>`,
-    ownership: 'United States of America'
-  }
-}
-function getOpenInvestmentProjects (investmentProjects) {
-  if (!investmentProjects) return null
-
-  return investmentProjects
-    .filter(project => project.open)
-    .map((project) => {
-      return {
-        name: `<a href="/investmentprojects/${project.id}/">${project.name}</a>`,
-        value: project.value,
-        state: project.state,
-        land_date: DateLib.formatDate(project.land_date)
-      }
-    })
-}
-function getClosedInvestmentProjects (investmentProjects) {
-  if (!investmentProjects) return null
-
-  return investmentProjects
-    .filter(project => !project.open)
-    .map((project) => {
-      return {
-        name: `<a href="/investmentprojects/${project.id}/">${project.name}</a>`,
-        value: project.value,
-        state: project.state,
-        state_date: DateLib.formatDate(project.state_date)
-      }
-    })
-}
 
 function index (req, res, next) {
   const id = req.params.sourceId
@@ -234,15 +44,15 @@ function index (req, res, next) {
     .then((company) => {
       const timeSinceNewContact = itemCollectionService.getTimeSinceLastAddedItem(company.contacts)
       const contactsInLastYear = itemCollectionService.getItemsAddedInLastYear(company.contacts)
-      const companyDisplay = getDisplayCompany(company)
-      const chDisplay = getDisplayCH(company)
-      const headingAddress = getHeadingAddress(company)
-      const headingName = getHeadingName(company)
-      const parents = parseRelatedData(company.parents)
-      const children = parseRelatedData(company.children)
-      const investmentDisplay = getInvestmentDetailsDisplay(company)
-      const investmentProjectsOpen = getOpenInvestmentProjects(company.investmentProjects)
-      const investmentProjectsClosed = getClosedInvestmentProjects(company.investmentProjects)
+      const companyDisplay = companyFormattingService.getDisplayCompany(company)
+      const chDisplay = companyFormattingService.getDisplayCH(company)
+      const headingAddress = companyFormattingService.getHeadingAddress(company)
+      const headingName = companyFormattingService.getHeadingName(company)
+      const parents = companyFormattingService.parseRelatedData(company.parents)
+      const children = companyFormattingService.parseRelatedData(company.children)
+      const investmentDisplay = investmentFormattingService.getInvestmentDetailsDisplay(company)
+      const investmentProjectsOpen = investmentFormattingService.getOpenInvestmentProjects(company.investmentProjects)
+      const investmentProjectsClosed = investmentFormattingService.getClosedInvestmentProjects(company.investmentProjects)
 
       res.render('company/index', {
         company,
@@ -251,23 +61,23 @@ function index (req, res, next) {
         timeSinceNewContact,
         contactsInLastYear,
         companyDetailLabels,
-        companyDetailsDisplayOrder,
+        companyDetailsDisplayOrder: Object.keys(companyDetailLabels),
         chDetailLabels,
-        chDetailsDisplayOrder,
+        chDetailsDisplayOrder: Object.keys(chDetailLabels),
         headingAddress,
         headingName,
         companyTableHeadings,
-        companyTableKeys,
+        companyTableKeys: Object.keys(companyTableHeadings),
         children,
         parents,
         investmentDetailLabels,
-        investmentDetailsDisplayOrder,
+        investmentDetailsDisplayOrder: Object.keys(investmentDetailLabels),
         investmentDisplay,
         investmentProjectsOpenLabels,
-        investmentProjectsOpenKeys,
+        investmentProjectsOpenKeys: Object.keys(investmentProjectsOpenLabels),
         investmentProjectsOpen,
         investmentProjectsClosedLabels,
-        investmentProjectsClosedKeys,
+        investmentProjectsClosedKeys: Object.keys(investmentProjectsClosedLabels),
         investmentProjectsClosed
       })
     })
@@ -357,4 +167,4 @@ router.post('/api/company', post)
 router.post('/api/company/archive', archive)
 router.post('/api/company/unarchive', unarchive)
 
-module.exports = { router, getDisplayCompany, getDisplayCH, getHeadingAddress }
+module.exports = { router }
