+++ src/labels/companylabels.js
@@ -1,0 +1,29 @@
+const companyDetailLabels = {
+  business_type: 'Business type',
+  name: 'Name',
+  registered_address: 'Primary address',
+  alias: 'Trading name',
+  trading_address: 'Trading address',
+  uk_region: 'UK region',
+  headquarters: 'Headquarters',
+  sector: 'Primary sector',
+  website: 'Website',
+  description: 'Business description',
+  employee_range: 'Number of employees',
+  turnover_range: 'Annual turnover'
+}
+const chDetailLabels = {
+  name: 'Registered company name',
+  company_number: 'Companies House number',
+  registered_address: 'Registered office address',
+  business_type: 'Company type',
+  company_status: 'Company status',
+  sic_code: 'Nature of business (SIC)'
+}
+const companyTableHeadings = {
+  name: 'Company name',
+  address: 'Address'
+}
+
+module.exports = { companyDetailLabels, chDetailLabels, companyTableHeadings }
+
