+++ src/labels/investmentlabels.js
@@ -1,0 +1,21 @@
+const investmentProjectsOpenLabels = {
+  name: 'Open projects',
+  value: 'Value',
+  state: 'Stage',
+  land_date: 'Land date'
+}
+
+const investmentProjectsClosedLabels = {
+  name: 'Closed projects',
+  value: 'Value',
+  state: 'Status',
+  state_date: 'Status date'
+}
+
+const investmentDetailLabels = {
+  account_management_tier: 'Account management tier',
+  account_manager: 'Account manager',
+  ownership: 'Ownership'
+}
+
+module.exports = { investmentProjectsOpenLabels, investmentProjectsClosedLabels, investmentDetailLabels }
