+++ src/sass/_shame.scss
@@ -93,6 +93,7 @@ hr {
   p.footnote {
     color: $grey-1;
     padding-left: 35px;
+    padding-bottom : 10px;
     margin: 0;
   }
 }
