+++ src/views/company/details.html
@@ -14,12 +14,20 @@
 
   <div class="section">
     <h3 class="heading-medium">Parent companies</h3>
-    {{  trade.datatable(parents, headings=companyTableHeadings, columnorder=companyTableKeys, class="table--related-companies") }}
+    {% if parents.length > 0 %}
+      {{  trade.datatable(parents, headings=companyTableHeadings, columnorder=companyTableKeys, class="table--related-companies") }}
+    {% else %}
+      <p>This company is currently not linked with any companies that it is a child of.</p>
+    {% endif %}
     <a href="#" class="button button-secondary">Add a parent company</a>
   </div>
   <div class="section">
     <h3 class="heading-medium">Child companies</h3>
-    {{  trade.datatable(children, headings=companyTableHeadings, columnorder=companyTableKeys, class="table--related-companies") }}
+    {% if children.length > 0 %}
+      {{  trade.datatable(children, headings=companyTableHeadings, columnorder=companyTableKeys, class="table--related-companies") }}
+    {% else %}
+      <p>This company is currently not linked with any companies that is is a parent of.</p>
+    {% endif %}
     <a href="#" class="button button-secondary">Add a child company</a>
   </div>
 {% else %}
